analyzer: split integer.h properly.
