+++ include/reaver/vapor/analyzer/types/integer.h
@@ -24,13 +24,9 @@
 
 #include <memory>
 
-#include <boost/multiprecision/integer.hpp>
-
-#include "../../codegen/ir/integer.h"
-#include "../../parser/literal.h"
-#include "../expressions/expression.h"
+#include "../../lexer/token.h"
 #include "../function.h"
-#include "../variables/literal.h"
+#include "type.h"
 
 namespace reaver::vapor::analyzer
 {
@@ -82,93 +78,6 @@ inline namespace _v1
         static function * _equal_comparison();
         static function * _less_comparison();
         static function * _less_equal_comparison();
-    };
-
-    class integer_constant : public literal
-    {
-    public:
-        integer_constant(const parser::integer_literal & parse) : _value{ utf8(parse.value.string) }
-        {
-        }
-
-        integer_constant(boost::multiprecision::cpp_int value) : _value{ std::move(value) }
-        {
-        }
-
-        virtual type * get_type() const override
-        {
-            return builtin_types().integer.get();
-        }
-
-        auto get_value() const
-        {
-            return _value;
-        }
-
-        virtual bool is_constant() const override
-        {
-            return true;
-        }
-
-        virtual bool is_equal(const variable * other_var) const override
-        {
-            auto other = dynamic_cast<const integer_constant *>(other_var);
-            if (!other)
-            {
-                // todo: conversions somehow
-                return false;
-            }
-
-            return other->get_value() == _value;
-        }
-
-    private:
-        virtual std::unique_ptr<variable> _clone_with_replacement(replacements &) const override
-        {
-            return std::make_unique<integer_constant>(_value);
-        }
-
-        virtual variable_ir _codegen_ir(ir_generation_context &) const override;
-
-        boost::multiprecision::cpp_int _value;
-    };
-
-    class integer_literal : public expression
-    {
-    public:
-        integer_literal(const parser::integer_literal & parse) : _parse{ parse }
-        {
-            auto val = std::make_unique<integer_constant>(parse);
-            _value = val.get();
-            _set_variable(std::move(val));
-        }
-
-        virtual void print(std::ostream & os, std::size_t indent) const override
-        {
-            auto in = std::string(indent, ' ');
-            os << in << "integer literal with value of " << _value->get_value() << " at " << _parse.range << '\n';
-        }
-
-    private:
-        virtual future<> _analyze(analysis_context &) override
-        {
-            return make_ready_future();
-        }
-
-        virtual std::unique_ptr<expression> _clone_expr_with_replacement(replacements & repl) const override
-        {
-            return make_variable_expression(_value->clone_with_replacement(repl));
-        }
-
-        virtual future<expression *> _simplify_expr(simplification_context &) override
-        {
-            return make_ready_future<expression *>(this);
-        }
-
-        virtual statement_ir _codegen_ir(ir_generation_context &) const override;
-
-        const parser::integer_literal & _parse;
-        integer_constant * _value;
     };
 
     std::unique_ptr<type> make_integer_type();
